Added benchmarks for the TemplateCompiler
micro-optimizations and disable DATAS to avoid skewing the performance
now that it is enabled by default in .NET 9.0
